azure_functions/upload_trigger/blob_upload_trigger/__init__.py
import logging

import azure.functions as func

from .processing import process


def main(myblob: func.InputStream):
    logging.info(
        f"Python blob trigger function processed blob \n"
        f"Name: {myblob.name.raw}\n"
        f"Blob Size: {myblob.length} bytes"
    )

    process(myblob.name.raw)
