azure_functions/upload_trigger/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure_storage==0.37.0
rasterio==1.3.5.post1
rio_cogeo==3.5.0